+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.4.0
+* Sanitize encrypted_card_data logs
+* Add `LocalPaymentExpired` and `LocalPaymentFunded` webhook notification support
+
 ## 4.3.0
 * Add a log message to the `ArgumentError` at `TransactionGateway.find`
 * Add `exchange_rate_quote_id` to `Transaction.create`